+++ src/SparseLUSolver.cpp
@@ -12,7 +12,7 @@ void SparseLUSolver::initialize(){
     // default Eigen representation: column major, which is good for klu !
     // J is const here
     auto timer = CustTimer();
-    n_ = J_.cols(); // should be equal to J_.nrows()
+    n_ = static_cast<int>(J_.cols()); // should be equal to J_.nrows()
     err_ = 0; // reset error message
     J_.makeCompressed();
     solver_.analyzePattern(J_);  //NEW
